% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hellox}
\alias{hellox}
\title{Show the CI/CD functionality}
\usage{
hellox(x)
}
\arguments{
\item{x}{The name of the Person who is being welcomed}
}
\value{
The output from \code{\link{print}}
}
\description{
Show the CI/CD functionality
}
\examples{
hellox("Zech")
\dontrun{
hellox("Rose")
}
}
